refactor(core): split tool_execution/tier_loop.rs into concern-scoped modules
tier_loop.rs mixed 8+ unrelated concerns (confirmation/escalation,
retry/reformat, quota, exfiltration checks, causal probing, focus
compression, shadow-sentinel events, doom-loop detection) in a single
3,340-line impl block, the largest non-test file in zeph-core.

Split along the same lines agent/state/ already uses: extract
confirmation.rs, retry_reformat.rs, quota.rs, exfiltration.rs,
causal_probe.rs, focus_compression.rs, shadow_events.rs, and
doom_loop.rs as sibling modules, keeping tier_loop.rs as the thin
orchestration entry point around run_tier_execution_loop.

Pure structural move: method bodies, tests, and visibility are
unchanged except cancel_tool_batch and skipped_output, widened from
private to pub(super) for their new cross-file callers.

Closes #6489
