+++ crates/zeph-core/src/agent/tool_execution/tier_loop.rs
@@ -12,16 +12,10 @@ use zeph_llm::provider::ToolDefinition;
 
 use super::{
     CacheCheckResult, TierLoopData, TierLoopOutput, ToolDispatchContext, ToolExecFut,
-    retry_backoff_ms, strip_tafc_fields, tool_args_hash,
+    strip_tafc_fields, tool_args_hash,
 };
 use crate::agent::Agent;
 use crate::channel::{Channel, StopHint, ToolStartEvent};
-
-/// Per-call timeout for the parameter-reformat LLM call (#5453) when
-/// `tools.retry.max_retry_duration_secs` is `0` ("no phase budget limit") — that value must not
-/// be reused directly as a per-call timeout (critic finding M1: `max(1)` on `0` collapsed to a
-/// 1-second timeout that always failed, silently disabling the whole feature).
-const REFORMAT_DEFAULT_TIMEOUT_SECS: u64 = 30;
 
 /// Build MAGE hard-block tier results when `is_blocked()` fired at dispatch time.
 ///
@@ -117,7 +111,10 @@ impl<C: Channel> Agent<C> {
     /// The tombstone persist runs before the notification send and its failure is logged
     /// rather than propagated: a closed/dropped channel receiver or disconnected adapter must
     /// never skip the tombstone and reintroduce the orphaned-`tool_calls` defect (#5717).
-    async fn cancel_tool_batch(
+    ///
+    /// `pub(super)`: called from the confirmation and retry/reformat phase modules after the
+    /// `tool_execution` split.
+    pub(super) async fn cancel_tool_batch(
         &mut self,
         tool_calls: &[zeph_llm::provider::ToolUseRequest],
         log_msg: &str,
@@ -133,392 +130,6 @@ impl<C: Channel> Agent<C> {
             );
         }
         Ok(())
-    }
-
-    /// Single batch-level human confirmation for the MAGE soft-escalation tier (spec 004-16
-    /// FR-006).
-    ///
-    /// Returns `Ok(true)` if the user declined — the tombstone and `[Cancelled]` notice are
-    /// already persisted via `cancel_tool_batch`, matching every other cancellation checkpoint
-    /// in this file; the caller must return `Ok(false)` without running the tier loop. Returns
-    /// `Ok(false)` if the user approved — the caller must then run the *normal*
-    /// `run_tier_execution_loop` so `check_trust`/`PermissionPolicy`/shadow-probe still apply
-    /// per call. MAGE escalation gates *whether* execution proceeds at all; it must never
-    /// substitute for those per-call gates — an earlier version of this wiring synthesized
-    /// `ToolError::ConfirmationRequired` and dispatched approved calls through
-    /// `execute_tool_call_confirmed_erased`, which explicitly skips `check_trust`, letting a
-    /// policy-`Deny` tool execute under escalation (critic finding F1).
-    async fn confirm_mage_escalation(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-    ) -> Result<bool, crate::agent::error::AgentError> {
-        let score = self.services.security.mage_accumulator.current_risk();
-        let prompt = format!(
-            "Elevated trajectory risk detected (score {score:.3}) — allow tool execution to proceed?"
-        );
-        if self.channel.confirm(&prompt).await? {
-            return Ok(false);
-        }
-        self.cancel_tool_batch(
-            tool_calls,
-            "tool execution cancelled: MAGE trajectory risk escalation declined",
-        )
-        .await?;
-        Ok(true)
-    }
-
-    #[tracing::instrument(
-        name = "core.tool.handle_confirmation_phase",
-        skip_all,
-        level = "debug",
-        err
-    )]
-    /// Returns `Ok(true)` if the user cancelled the turn during this phase.
-    async fn handle_confirmation_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<bool, crate::agent::error::AgentError> {
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.cancel_tool_batch(tool_calls, "tool execution cancelled by user")
-                    .await?;
-                return Ok(true);
-            }
-            let new_result =
-                if let Err(zeph_tools::ToolError::ConfirmationRequired { ref command }) =
-                    tool_results[idx]
-                {
-                    let tc = &tool_calls[idx];
-                    let prompt = if command.is_empty() {
-                        format!("Allow tool: {}?", tc.name)
-                    } else {
-                        format!("Allow command: {command}?")
-                    };
-                    Some(if self.channel.confirm(&prompt).await? {
-                        // execute_tool_call_confirmed_erased bypasses check_trust; a second
-                        // ConfirmationRequired here indicates a misconfigured executor stack.
-                        self.tool_executor
-                            .execute_tool_call_confirmed_erased(&calls[idx])
-                            .await
-                    } else {
-                        Ok(Some(zeph_tools::ToolOutput {
-                            tool_name: tc.name.clone(),
-                            summary: "[cancelled by user]".to_owned(),
-                            blocks_executed: 0,
-                            filter_stats: None,
-                            diff: None,
-                            streamed: false,
-                            terminal_id: None,
-                            locations: None,
-                            raw_response: None,
-                            claim_source: None,
-                            ..Default::default()
-                        }))
-                    })
-                } else {
-                    None
-                };
-            if let Some(result) = new_result {
-                if let Err(ref e) = result
-                    && let Some(ref d) = self.runtime.debug.debug_dumper
-                {
-                    d.dump_tool_error(tool_calls[idx].name.as_str(), e);
-                }
-                tool_results[idx] = result;
-            }
-        }
-        Ok(false)
-    }
-
-    /// Returns `Ok(true)` if the user cancelled the turn during this phase.
-    #[tracing::instrument(name = "core.tool.handle_retry_phase", skip_all, level = "debug", err)]
-    async fn handle_retry_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        max_retries: usize,
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<bool, crate::agent::error::AgentError> {
-        if max_retries == 0 {
-            return Ok(false);
-        }
-        let max_retry_duration_secs = self.tool_orchestrator.max_retry_duration_secs;
-        let retry_base_ms = self.tool_orchestrator.retry_base_ms;
-        let retry_max_ms = self.tool_orchestrator.retry_max_ms;
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.cancel_tool_batch(tool_calls, "tool execution cancelled by user")
-                    .await?;
-                return Ok(true);
-            }
-            let is_transient = matches!(
-                tool_results[idx],
-                Err(ref e) if e.kind() == zeph_tools::ErrorKind::Transient
-            );
-            if !is_transient {
-                continue;
-            }
-            let tc = &tool_calls[idx];
-            if !self
-                .tool_executor
-                .is_tool_retryable_erased(tc.name.as_str())
-            {
-                continue;
-            }
-            let call = &calls[idx];
-            let mut attempt = 0_usize;
-            let retry_start = std::time::Instant::now();
-            let result = loop {
-                let exec_result = tokio::select! {
-                    r = self.tool_executor.execute_tool_call_erased(call).instrument(
-                        tracing::info_span!("tool_exec_retry", tool_name = %tc.name, idx = %tc.id)
-                    ) => r,
-                    () = cancel.cancelled() => {
-                        self.cancel_tool_batch(tool_calls, "tool retry cancelled by user")
-                            .await?;
-                        return Ok(true);
-                    }
-                };
-                match exec_result {
-                    Err(ref e)
-                        if e.kind() == zeph_tools::ErrorKind::Transient
-                            && attempt < max_retries =>
-                    {
-                        let elapsed_secs = retry_start.elapsed().as_secs();
-                        if max_retry_duration_secs > 0 && elapsed_secs >= max_retry_duration_secs {
-                            tracing::warn!(
-                                tool = %tc.name, elapsed_secs, max_retry_duration_secs,
-                                "tool retry budget exceeded, aborting retries"
-                            );
-                            break exec_result;
-                        }
-                        attempt += 1;
-                        let delay_ms = retry_backoff_ms(attempt - 1, retry_base_ms, retry_max_ms);
-                        tracing::warn!(
-                            tool = %tc.name, attempt, delay_ms, error = %e,
-                            "transient tool error, retrying with backoff"
-                        );
-                        self.channel
-                            .send_status_best_effort(&format!("Retrying {}...", tc.name))
-                            .await;
-                        // Interruptible backoff sleep: cancelled if agent shuts down.
-                        tokio::select! {
-                            () = tokio::time::sleep(std::time::Duration::from_millis(delay_ms)) => {}
-                            () = cancel.cancelled() => {
-                                self.cancel_tool_batch(
-                                    tool_calls,
-                                    "retry backoff interrupted by cancellation",
-                                )
-                                .await?;
-                                return Ok(true);
-                            }
-                        }
-                        self.channel.send_status_best_effort("").await;
-                        // NOTE: retry re-executions are NOT recorded in repeat-detection (CRIT-3).
-                    }
-                    result => break result,
-                }
-            };
-            tool_results[idx] = result;
-        }
-        Ok(false)
-    }
-
-    /// Returns `Ok(true)` if the user cancelled the turn during this phase.
-    #[tracing::instrument(
-        name = "core.tool.handle_reformat_phase",
-        skip_all,
-        level = "debug",
-        err
-    )]
-    async fn handle_reformat_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<bool, crate::agent::error::AgentError> {
-        if self
-            .tool_orchestrator
-            .parameter_reformat_provider
-            .is_empty()
-        {
-            return Ok(false);
-        }
-        // Budget covers the whole reformat phase (all tool calls needing reformat this round),
-        // matching the retry phase's `retry_start`/`max_retry_duration_secs` accounting.
-        let budget_secs = self.tool_orchestrator.max_retry_duration_secs;
-        let reformat_start = std::time::Instant::now();
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.cancel_tool_batch(tool_calls, "parameter reformat phase cancelled by user")
-                    .await?;
-                return Ok(true);
-            }
-            let needs_reformat = matches!(
-                tool_results[idx],
-                Err(ref e) if e.category().needs_parameter_reformat()
-            );
-            if !needs_reformat {
-                continue;
-            }
-            let tc = &tool_calls[idx];
-            if budget_secs > 0 && reformat_start.elapsed().as_secs() >= budget_secs {
-                tracing::warn!(tool = %tc.name, "parameter reformat budget exhausted, skipping");
-                continue;
-            }
-            let error_message = tool_results[idx]
-                .as_ref()
-                .err()
-                .map(std::string::ToString::to_string)
-                .unwrap_or_default();
-
-            self.channel
-                .send_status_best_effort(&format!("Reformatting parameters for {}...", tc.name))
-                .await;
-            let new_result = self
-                .reformat_tool_call(&calls[idx], tc, &error_message, cancel)
-                .await;
-            self.channel.send_status_best_effort("").await;
-
-            if cancel.is_cancelled() {
-                self.cancel_tool_batch(tool_calls, "parameter reformat phase cancelled by user")
-                    .await?;
-                return Ok(true);
-            }
-
-            if let Some(result) = new_result {
-                if let Err(ref e) = result
-                    && let Some(ref d) = self.runtime.debug.debug_dumper
-                {
-                    d.dump_tool_error(tc.name.as_str(), e);
-                }
-                tool_results[idx] = result;
-            }
-        }
-        Ok(false)
-    }
-
-    /// LLM-based single-shot reformat-and-retry for a tool call that failed with an
-    /// `InvalidParameters`/`TypeMismatch` error (issue #5453).
-    ///
-    /// Resolves `tools.retry.parameter_reformat_provider` from `[[llm.providers]]` via
-    /// [`Agent::resolve_pool_entry_provider`]. Unlike most other background-provider call sites
-    /// (which use [`Agent::resolve_background_provider`] and fall back to the primary provider
-    /// on any resolution failure), a *configured* name — one the provider-pool registry was
-    /// actually wired to recognize — must not silently substitute the primary provider when it
-    /// fails to resolve: that would mask the original tool error behind a "corrected" call made
-    /// with the wrong model (#5600, #5478). This method no-ops instead (keeps the original tool
-    /// error) whenever the registry is wired but the name is absent from `provider_pool`, the
-    /// matched entry fails to build, or no `provider_config_snapshot` is available. It only
-    /// falls back to [`Agent::resolve_background_provider`]'s legacy convention when the
-    /// provider-pool registry itself was never wired for this `Agent` at all — which
-    /// `zeph_config::providers::validate_pool` guarantees cannot happen for a real, fully
-    /// constructed production agent (see [`Agent::resolve_pool_entry_provider`]'s doc comment).
-    ///
-    /// Returns `None` when the provider is unresolvable, the provider call fails, times out, or
-    /// returns arguments that do not parse as a JSON object — the caller keeps the original error
-    /// result unchanged.
-    async fn reformat_tool_call(
-        &mut self,
-        call: &ToolCall,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        error_message: &str,
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Option<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>> {
-        #[derive(Debug, serde::Deserialize, schemars::JsonSchema)]
-        struct ReformattedArguments {
-            /// Corrected JSON arguments object for the failed tool call.
-            arguments: serde_json::Value,
-        }
-
-        let Some(schema) = self
-            .tool_executor
-            .tool_definitions_erased()
-            .into_iter()
-            .find(|d| d.id.as_ref() == tc.name.as_str())
-            .map(|d| d.schema)
-        else {
-            tracing::warn!(tool = %tc.name, "parameter reformat: tool schema not found, skipping");
-            return None;
-        };
-
-        let provider_name = self.tool_orchestrator.parameter_reformat_provider.clone();
-        let provider = match self.resolve_pool_entry_provider(&provider_name) {
-            super::super::learning::PoolProviderResolution::Resolved(p) => *p,
-            super::super::learning::PoolProviderResolution::RegistryNotWired => {
-                self.resolve_background_provider(&provider_name)
-            }
-            super::super::learning::PoolProviderResolution::Unresolvable => {
-                tracing::warn!(
-                    tool = %tc.name,
-                    provider = %provider_name,
-                    "parameter reformat: configured provider unresolvable, keeping original error"
-                );
-                return None;
-            }
-        };
-
-        let original_args = serde_json::Value::Object(call.params.clone());
-        let prompt = format!(
-            "A tool call failed parameter validation. Propose corrected arguments as a JSON \
-             object under the `arguments` key.\n\n\
-             Tool: {}\nJSON schema:\n{}\n\nOriginal arguments:\n{}\n\nError: {error_message}",
-            tc.name,
-            serde_json::to_string_pretty(&schema).unwrap_or_default(),
-            serde_json::to_string_pretty(&original_args).unwrap_or_default(),
-        );
-        let messages = [Message::from_legacy(Role::User, prompt)];
-
-        // `max_retry_duration_secs == 0` means "no phase budget limit" (see the `budget_secs > 0`
-        // check in `handle_reformat_phase`), but a per-LLM-call timeout must never be 0 or
-        // collapse to 1s for that same value — that previously made `max_retry_duration_secs = 0`
-        // ("unlimited") silently disable the whole feature via an always-timing-out 1s call
-        // (critic finding M1). Use the configured budget as the per-call timeout only when it is a
-        // real bound; fall back to a fixed sane default otherwise.
-        let timeout_secs = match self.tool_orchestrator.max_retry_duration_secs {
-            0 => REFORMAT_DEFAULT_TIMEOUT_SECS,
-            secs => secs,
-        };
-        let reformat = tokio::select! {
-            r = tokio::time::timeout(
-                std::time::Duration::from_secs(timeout_secs),
-                provider.chat_typed_erased::<ReformattedArguments>(&messages),
-            ) => match r {
-                Ok(Ok(reformat)) => reformat,
-                Ok(Err(e)) => {
-                    tracing::warn!(tool = %tc.name, error = %e, "parameter reformat: provider call failed");
-                    return None;
-                }
-                Err(_) => {
-                    tracing::warn!(tool = %tc.name, timeout_secs, "parameter reformat: provider call timed out");
-                    return None;
-                }
-            },
-            () = cancel.cancelled() => return None,
-        };
-
-        let serde_json::Value::Object(corrected_params) = reformat.arguments else {
-            tracing::warn!(
-                tool = %tc.name,
-                "parameter reformat: corrected arguments were not a JSON object, skipping retry"
-            );
-            return None;
-        };
-
-        let mut retry_call = call.clone();
-        retry_call.params = corrected_params;
-
-        tokio::select! {
-            r = self.tool_executor.execute_tool_call_erased(&retry_call).instrument(
-                tracing::info_span!("tool_exec_reformat", tool_name = %tc.name)
-            ) => Some(r),
-            () = cancel.cancelled() => None,
-        }
     }
 
     fn run_pre_execution_verifiers(&mut self, calls: &[ToolCall]) -> Vec<bool> {
@@ -1138,133 +749,6 @@ impl<C: Channel> Agent<C> {
         }
     }
 
-    /// Check MAGE trajectory risk gate (spec 004-16 FR-004, FR-005).
-    ///
-    /// Returns `Some((score, top_signals))` when the accumulator is blocked. Emits a security
-    /// event, increments `pre_execution_blocks`, and calls `record_block()` on the accumulator.
-    fn check_mage_block(&mut self) -> Option<(f64, Vec<String>)> {
-        if !self.services.security.mage_accumulator.is_blocked() {
-            return None;
-        }
-        let score = self.services.security.mage_accumulator.current_risk();
-        let top: Vec<String> = self
-            .services
-            .security
-            .mage_accumulator
-            .top_signals(3)
-            .iter()
-            .map(|s| format!("{:?}({:?})", s.signal_type, s.severity))
-            .collect();
-        tracing::warn!(
-            score,
-            signals = ?top,
-            "MAGE trajectory risk accumulator blocked tool dispatch"
-        );
-        self.update_metrics(|m| m.pre_execution_blocks += 1);
-        self.push_security_event(
-            zeph_common::SecurityEventCategory::PreExecutionBlock,
-            "<mage>",
-            format!("trajectory risk {score:.3} exceeds threshold"),
-        );
-        self.services.security.mage_accumulator.record_block();
-        Some((score, top))
-    }
-
-    /// Check MAGE trajectory risk soft-escalation gate (spec 004-16 FR-006).
-    ///
-    /// Returns `true` when the accumulator's risk is in `[escalation_threshold,
-    /// risk_threshold)`. Emits a security event, increments `pre_execution_warnings`, and
-    /// calls `record_escalation()` on the accumulator so the caller gates the batch behind
-    /// a single `Agent::confirm_mage_escalation` confirmation before falling through to the
-    /// normal tier execution loop (see that method's doc comment for why this must not
-    /// bypass `check_trust`/`PermissionPolicy`).
-    fn check_mage_escalation(&mut self) -> bool {
-        if !self.services.security.mage_accumulator.should_escalate() {
-            return false;
-        }
-        let score = self.services.security.mage_accumulator.current_risk();
-        tracing::warn!(
-            score,
-            "MAGE trajectory risk accumulator escalating tool dispatch to human confirmation"
-        );
-        self.update_metrics(|m| m.pre_execution_warnings += 1);
-        self.push_security_event(
-            zeph_common::SecurityEventCategory::PreExecutionWarn,
-            "<mage>",
-            format!("trajectory risk {score:.3} in escalation band, requiring confirmation"),
-        );
-        self.services.security.mage_accumulator.record_escalation();
-        true
-    }
-
-    fn check_and_update_quota(&mut self, batch_len: usize) -> bool {
-        if let Some(max) = self.tool_orchestrator.check_quota() {
-            tracing::warn!(
-                max,
-                count = self.tool_orchestrator.session_tool_call_count,
-                "tool call quota exceeded for session"
-            );
-            return true;
-        }
-        let batch_count = u32::try_from(batch_len).unwrap_or(u32::MAX);
-        self.tool_orchestrator.session_tool_call_count = self
-            .tool_orchestrator
-            .session_tool_call_count
-            .saturating_add(batch_count);
-        self.runtime.lifecycle.turn_tool_calls = self
-            .runtime
-            .lifecycle
-            .turn_tool_calls
-            .saturating_add(batch_count);
-        false
-    }
-
-    fn check_exfiltration_urls(&mut self, tool_calls: &[zeph_llm::provider::ToolUseRequest]) {
-        for tc in tool_calls {
-            let args_json = tc.input.to_string();
-            let url_events = self
-                .services
-                .security
-                .exfiltration_guard
-                .validate_tool_call(
-                    tc.name.as_str(),
-                    &args_json,
-                    &self.services.security.flagged_urls,
-                );
-            if !url_events.is_empty() {
-                tracing::warn!(
-                    tool = %tc.name,
-                    count = url_events.len(),
-                    "exfiltration guard: suspicious URLs in tool arguments (flag-only, not blocked)"
-                );
-                self.update_metrics(|m| {
-                    m.exfiltration_tool_urls_flagged += url_events.len() as u64;
-                });
-                self.push_security_event(
-                    zeph_common::SecurityEventCategory::ExfiltrationBlock,
-                    tc.name.as_str(),
-                    format!(
-                        "{} suspicious URL(s) flagged in tool args",
-                        url_events.len()
-                    ),
-                );
-            }
-        }
-    }
-
-    #[tracing::instrument(name = "core.tool.run_causal_pre_probe", skip_all, level = "debug")]
-    async fn run_causal_pre_probe(&mut self) -> Option<(String, String)> {
-        let analyzer = self.services.security.causal_analyzer.as_ref()?;
-        let context_summary = self.build_causal_context_summary();
-        match analyzer.probe(&context_summary).await {
-            Ok(resp) => Some((resp, context_summary)),
-            Err(e) => {
-                tracing::warn!(error = %e, "causal IPI pre-probe failed, skipping analysis");
-                None
-            }
-        }
-    }
-
     #[tracing::instrument(
         name = "core.tool.run_tier_execution_loop",
         skip_all,
@@ -1457,47 +941,6 @@ impl<C: Channel> Agent<C> {
             pending_focus_checkpoint,
             pending_system_hints,
         }))
-    }
-
-    #[tracing::instrument(
-        name = "core.tool.preprocess_focus_compress",
-        skip_all,
-        level = "debug"
-    )]
-    async fn preprocess_focus_compress_calls(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-    ) -> Option<zeph_llm::provider::Message> {
-        let mut pending_focus_checkpoint: Option<zeph_llm::provider::Message> = None;
-        for (idx, tc) in tool_calls.iter().enumerate() {
-            let is_focus_tool = self.services.focus.config.enabled
-                && (tc.name == "start_focus" || tc.name == "complete_focus");
-            let is_compress = tc.name == "compress_context";
-            let is_request_compaction = tc.name == "request_compaction"
-                && self
-                    .services
-                    .memory
-                    .subsystems
-                    .arc_config
-                    .allow_agent_compaction;
-            if is_focus_tool || is_compress || is_request_compaction {
-                let result = if is_compress {
-                    self.handle_compress_context().await
-                } else if is_request_compaction {
-                    self.handle_request_compaction(&tc.input).await
-                } else {
-                    let (text, maybe_checkpoint) =
-                        self.handle_focus_tool(tc.name.as_str(), &tc.input);
-                    if let Some(cp) = maybe_checkpoint {
-                        pending_focus_checkpoint = Some(cp);
-                    }
-                    text
-                };
-                tool_results[idx] = Ok(Some(skipped_output(tc.name.clone(), result)));
-            }
-        }
-        pending_focus_checkpoint
     }
 
     pub(super) async fn stamp_and_send_tier_start(
@@ -1912,45 +1355,6 @@ impl<C: Channel> Agent<C> {
             )));
         }
         Ok(None)
-    }
-
-    /// Fires `permission_denied` hooks (fail-open). Called at every gate/rate-limiter denial.
-    ///
-    /// Hooks run sequentially; slow or hanging hooks will stall tool dispatch for each denied
-    /// call. Hook authors should ensure hooks complete quickly or use a background process.
-    async fn fire_permission_denied_hooks(
-        &mut self,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        reason: &str,
-    ) {
-        let pd_hooks = self.services.session.hooks_config.permission_denied.clone();
-        if pd_hooks.is_empty() {
-            return;
-        }
-        let mut env = std::collections::HashMap::new();
-        env.insert("ZEPH_DENIED_TOOL".to_owned(), tc.name.to_string());
-        env.insert("ZEPH_DENY_REASON".to_owned(), reason.to_owned());
-        env.insert("ZEPH_TOOL_NAME".to_owned(), tc.name.to_string());
-        let conv_id_str = self
-            .services
-            .memory
-            .persistence
-            .conversation_id
-            .map(|id| id.0.to_string());
-        crate::agent::hooks_dispatch::insert_main_agent_ctx(&mut env, conv_id_str.as_deref());
-        let dispatch = self.mcp_dispatch();
-        let mcp: Option<&dyn zeph_subagent::McpDispatch> = dispatch
-            .as_ref()
-            .map(|d| d as &dyn zeph_subagent::McpDispatch);
-        if let Err(e) = zeph_subagent::hooks::fire_hooks(&pd_hooks, &env, mcp, None)
-            .instrument(tracing::info_span!(
-                "core.hooks.permission_denied",
-                tool = %tc.name
-            ))
-            .await
-        {
-            tracing::warn!(error = %e, tool = %tc.name, "PermissionDenied hook failed");
-        }
     }
 
     #[tracing::instrument(
@@ -2457,71 +1861,6 @@ impl<C: Channel> Agent<C> {
     }
 
     #[tracing::instrument(
-        name = "core.tool.run_causal_ipi_post_probe",
-        skip_all,
-        level = "debug"
-    )]
-    async fn run_causal_ipi_post_probe(
-        &mut self,
-        causal_pre_response: Option<(String, String)>,
-        result_parts: &[MessagePart],
-    ) {
-        let Some((pre_response, context_summary)) = causal_pre_response else {
-            return;
-        };
-        let snippets: Vec<String> = result_parts
-            .iter()
-            .filter_map(|p| {
-                if let MessagePart::ToolResult {
-                    content, is_error, ..
-                } = p
-                {
-                    if *is_error {
-                        Some(zeph_sanitizer::causal_ipi::format_error_snippet(content))
-                    } else {
-                        Some(zeph_sanitizer::causal_ipi::format_tool_snippet(content))
-                    }
-                } else {
-                    None
-                }
-            })
-            .collect();
-        let tool_snippets = if snippets.is_empty() {
-            "[empty]".to_owned()
-        } else {
-            snippets.join("---")
-        };
-        let Some(ref analyzer) = self.services.security.causal_analyzer else {
-            return;
-        };
-        match analyzer.post_probe(&context_summary, &tool_snippets).await {
-            Ok(post_response) => {
-                let analysis = analyzer.analyze(&pre_response, &post_response);
-                if analysis.is_flagged {
-                    let pre_excerpt = &pre_response[..pre_response.floor_char_boundary(100)];
-                    let post_excerpt = &post_response[..post_response.floor_char_boundary(100)];
-                    tracing::warn!(
-                        deviation_score = analysis.deviation_score,
-                        threshold = analyzer.threshold(),
-                        pre = %pre_excerpt,
-                        post = %post_excerpt,
-                        "causal IPI: behavioral deviation detected at tool-return boundary"
-                    );
-                    self.update_metrics(|m| m.causal_ipi_flags += 1);
-                    self.push_security_event(
-                        zeph_common::SecurityEventCategory::CausalIpiFlag,
-                        "tool_batch",
-                        format!("deviation={:.3}", analysis.deviation_score),
-                    );
-                }
-            }
-            Err(e) => {
-                tracing::warn!(error = %e, "causal IPI post-probe failed, skipping analysis");
-            }
-        }
-    }
-
-    #[tracing::instrument(
         name = "core.tool.process_tool_result_batch",
         skip_all,
         level = "debug",
@@ -2723,154 +2062,6 @@ impl<C: Channel> Agent<C> {
 
         Ok(())
     }
-
-    /// Record a [`ShadowEvent`](zeph_sanitizer::ShadowEvent) for cross-turn goal-drift detection.
-    ///
-    /// Called after every tool batch completes (spec 010-7 FR-001). When `shadow_memory` is
-    /// `None` (disabled) this is a no-op. When drift score triggers an alert, emits a
-    /// `GoalDrift` security event (FR-003, FR-004).
-    fn record_shadow_event(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        goal_summary: String,
-    ) {
-        let Some(ref mut mem) = self.services.security.shadow_memory else {
-            return;
-        };
-        let tool_names: Vec<String> = tool_calls
-            .iter()
-            .map(|tc| tc.name.as_str().to_owned())
-            .collect();
-        let max_permission_class = tool_names
-            .iter()
-            .map(|n| zeph_sanitizer::classify_tool_permission(n))
-            .max()
-            .unwrap_or(0);
-        let turn = u32::try_from(self.runtime.debug.iteration_counter.saturating_sub(1))
-            .unwrap_or(u32::MAX);
-        mem.record(zeph_sanitizer::ShadowEvent {
-            turn,
-            tools: tool_names,
-            max_permission_class,
-            deviation_score: 0.0,
-            goal_summary,
-        });
-        let drift = mem.goal_drift_score();
-        if drift.should_alert {
-            tracing::warn!(
-                score = drift.score,
-                turn = turn,
-                "shadow memory: goal drift alert"
-            );
-            self.push_security_event(
-                zeph_common::SecurityEventCategory::GoalDrift,
-                "shadow_memory",
-                format!("drift_score={:.3}", drift.score),
-            );
-        }
-    }
-
-    /// Apply Acon tool-result compression to `result_parts` in-place before the parts enter
-    /// message history. No-op when `acon_config.enabled` is false or the batch is empty.
-    #[tracing::instrument(name = "context.tool_result_compress", skip_all, level = "debug")]
-    fn apply_acon_compression(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        result_parts: &mut [MessagePart],
-    ) {
-        use zeph_context::tool_result_compress::{
-            CompressionMethod, ToolResultCompressionConfig, ToolResultCompressor, ToolResultEntry,
-        };
-
-        let acon = &self.services.memory.subsystems.acon_config;
-        if !acon.enabled {
-            return;
-        }
-
-        let cfg = ToolResultCompressionConfig::from(acon);
-        let tc = std::sync::Arc::clone(&self.runtime.metrics.token_counter);
-
-        // Build a lookup from tool_use_id → tool_name so we can populate the trace field
-        // without relying on positional correspondence between result_parts and tool_calls.
-        // This is robust to future changes where process_one_tool_result emits zero or
-        // multiple ToolResult parts per call.
-        let id_to_name: std::collections::HashMap<&str, &str> = tool_calls
-            .iter()
-            .map(|tc| (tc.id.as_str(), tc.name.as_str()))
-            .collect();
-
-        // Collect (part_index, tool_name, owned_text) for each ToolResult part. Text is cloned
-        // to avoid borrow conflicts when we later mutate result_parts.
-        let indexed_texts: Vec<(usize, String, String)> = result_parts
-            .iter()
-            .enumerate()
-            .filter_map(|(i, part)| {
-                if let MessagePart::ToolResult {
-                    content,
-                    tool_use_id,
-                    ..
-                } = part
-                {
-                    let name = id_to_name
-                        .get(tool_use_id.as_str())
-                        .copied()
-                        .unwrap_or("")
-                        .to_owned();
-                    Some((i, name, content.clone()))
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        if indexed_texts.is_empty() {
-            return;
-        }
-
-        let entries: Vec<ToolResultEntry<'_>> = indexed_texts
-            .iter()
-            .map(|(part_idx, name, text)| ToolResultEntry {
-                tool_name: name.as_str(),
-                text: text.as_str(),
-                index: *part_idx,
-            })
-            .collect();
-
-        let compressed = ToolResultCompressor::compress_batch(&entries, tc.as_ref(), &cfg);
-
-        let mut tokens_saved: usize = 0;
-        let mut results_compressed: u32 = 0;
-
-        for (result, (part_idx, _, _)) in compressed.iter().zip(indexed_texts.iter()) {
-            if result.method != CompressionMethod::PassThrough
-                && let MessagePart::ToolResult { content, .. } = &mut result_parts[*part_idx]
-            {
-                content.clone_from(&result.text);
-                tokens_saved = tokens_saved.saturating_add(
-                    result
-                        .original_tokens
-                        .saturating_sub(result.compressed_tokens),
-                );
-                results_compressed += 1;
-            }
-        }
-
-        if results_compressed > 0 {
-            tracing::debug!(
-                tokens_saved,
-                results_compressed,
-                "acon: tool result compression applied"
-            );
-            self.update_metrics(|m| {
-                m.acon_tokens_saved = m
-                    .acon_tokens_saved
-                    .saturating_add(u64::try_from(tokens_saved).unwrap_or(u64::MAX));
-                m.acon_results_compressed = m
-                    .acon_results_compressed
-                    .saturating_add(u64::from(results_compressed));
-            });
-        }
-    }
 }
 
 /// Recursively unmask PAAC secret placeholders in a JSON value's string leaves, in place.
@@ -2920,7 +2111,8 @@ async fn recv_elicitation(
     }
 }
 
-fn skipped_output(
+// `pub(super)`: called from `focus_compression` after the `tool_execution` split.
+pub(super) fn skipped_output(
     tool_name: impl Into<zeph_common::ToolName>,
     summary: impl Into<String>,
 ) -> zeph_tools::ToolOutput {
@@ -3059,39 +2251,6 @@ impl<C: Channel> Agent<C> {
         let _ = self.channel.send_stop_hint(StopHint::MaxTurnRequests).await;
         self.channel.flush_chunks().await?;
         Ok(())
-    }
-
-    /// Returns `true` if a doom loop was detected and the caller should break.
-    async fn check_doom_loop(
-        &mut self,
-        iteration: usize,
-    ) -> Result<bool, crate::agent::error::AgentError> {
-        if let Some(last_msg) = self.msg.messages.last() {
-            let hash = zeph_agent_tools::doom_loop_hash(&last_msg.content);
-            tracing::debug!(
-                iteration,
-                hash,
-                content_len = last_msg.content.len(),
-                content_preview = &last_msg.content[..last_msg.content.len().min(120)],
-                "doom-loop hash recorded"
-            );
-            self.tool_orchestrator.push_doom_hash(hash);
-            if self.tool_orchestrator.is_doom_loop() {
-                tracing::warn!(
-                    iteration,
-                    hash,
-                    content_len = last_msg.content.len(),
-                    content_preview = &last_msg.content[..last_msg.content.len().min(200)],
-                    "doom-loop detected: {} consecutive identical outputs",
-                    crate::agent::DOOM_LOOP_WINDOW
-                );
-                self.channel
-                    .send("Stopping: detected repeated identical tool outputs.")
-                    .await?;
-                return Ok(true);
-            }
-        }
-        Ok(false)
     }
 
     /// Drive an LLM chat call through the durable step journal when a [`DurableContext`] is
@@ -3342,6 +2501,21 @@ mod tests {
         serde_json::from_str(s).unwrap()
     }
 
+    fn make_tool_req(name: &str) -> zeph_llm::provider::ToolUseRequest {
+        zeph_llm::provider::ToolUseRequest {
+            id: format!("id_{name}"),
+            name: name.into(),
+            input: serde_json::Value::Null,
+        }
+    }
+
+    fn make_tool_call(name: &str) -> ToolCall {
+        ToolCall {
+            tool_id: zeph_common::ToolName::new(name),
+            ..Default::default()
+        }
+    }
+
     #[test]
     fn make_tool_hook_env_sets_tool_name() {
         let env = make_tool_hook_env("Edit", &serde_json::Value::Null, None);
@@ -3441,238 +2615,6 @@ mod tests {
         assert!(matching_hooks(&matchers, "complete_focus").is_empty());
     }
 
-    // Regression guard for issue #3774: permission_denied hook env must contain
-    // ZEPH_DENIED_TOOL and ZEPH_DENY_REASON for every gate/rate-limiter denial.
-    // These tests verify the env construction logic mirrored in fire_permission_denied_hooks.
-
-    fn make_pd_env(tool: &str, reason: &str) -> std::collections::HashMap<String, String> {
-        let mut env = std::collections::HashMap::new();
-        env.insert("ZEPH_DENIED_TOOL".to_owned(), tool.to_owned());
-        env.insert("ZEPH_DENY_REASON".to_owned(), reason.to_owned());
-        env
-    }
-
-    #[test]
-    fn permission_denied_env_contains_tool_name_and_reason_for_quota_denial() {
-        let tool = "shell";
-        let reason = "session tool call quota exceeded (limit: 10 calls)";
-        let env = make_pd_env(tool, reason);
-
-        assert_eq!(
-            env.get("ZEPH_DENIED_TOOL").map(String::as_str),
-            Some("shell")
-        );
-        assert!(
-            env.get("ZEPH_DENY_REASON")
-                .is_some_and(|r| r.contains("quota")),
-            "ZEPH_DENY_REASON should mention quota"
-        );
-    }
-
-    #[test]
-    fn permission_denied_env_contains_tool_name_and_reason_for_rate_limit_denial() {
-        use crate::agent::rate_limiter::{RateLimitExceeded, ToolCategory};
-
-        let exceeded = RateLimitExceeded {
-            category: ToolCategory::Shell,
-            count: 5,
-            limit: 3,
-            cooldown_remaining_secs: 30,
-        };
-        let reason = exceeded.to_error_message();
-        let env = make_pd_env("bash", &reason);
-
-        assert_eq!(
-            env.get("ZEPH_DENIED_TOOL").map(String::as_str),
-            Some("bash")
-        );
-        let deny_reason = env
-            .get("ZEPH_DENY_REASON")
-            .expect("ZEPH_DENY_REASON missing");
-        assert!(
-            deny_reason.contains("rate-limited"),
-            "ZEPH_DENY_REASON should mention rate-limited, got: {deny_reason}"
-        );
-        assert!(
-            deny_reason.contains("3/min"),
-            "ZEPH_DENY_REASON should contain limit, got: {deny_reason}"
-        );
-    }
-
-    #[test]
-    fn permission_denied_env_contains_tool_name_and_reason_for_pre_exec_block() {
-        let tool = "write";
-        let reason = format!("blocked by pre-execution verifier: {tool} is not permitted");
-        let env = make_pd_env(tool, &reason);
-
-        assert_eq!(
-            env.get("ZEPH_DENIED_TOOL").map(String::as_str),
-            Some("write")
-        );
-        assert!(
-            env.get("ZEPH_DENY_REASON")
-                .is_some_and(|r| r.contains("pre-execution verifier")),
-            "ZEPH_DENY_REASON should mention pre-execution verifier"
-        );
-    }
-
-    #[test]
-    fn permission_denied_env_contains_tool_name_and_reason_for_repeat_block() {
-        let tool = "read";
-        let reason = format!("repeated identical call to {tool} detected");
-        let env = make_pd_env(tool, &reason);
-
-        assert_eq!(
-            env.get("ZEPH_DENIED_TOOL").map(String::as_str),
-            Some("read")
-        );
-        assert!(
-            env.get("ZEPH_DENY_REASON")
-                .is_some_and(|r| r.contains("repeated identical call")),
-            "ZEPH_DENY_REASON should mention repeated identical call"
-        );
-    }
-
-    #[test]
-    fn permission_denied_env_reason_includes_utility_action_variant() {
-        // Verify that utility gate reason strings include the UtilityAction Debug variant name
-        // so hook authors can distinguish Respond/Retrieve/Verify/Stop in ZEPH_DENY_REASON.
-        use zeph_tools::UtilityAction;
-
-        for action in [
-            UtilityAction::Respond,
-            UtilityAction::Retrieve,
-            UtilityAction::Verify,
-            UtilityAction::Stop,
-        ] {
-            let reason = format!("utility gate ({action:?}) intercepted memory_search");
-            let env = make_pd_env("memory_search", &reason);
-
-            let deny_reason = env
-                .get("ZEPH_DENY_REASON")
-                .expect("ZEPH_DENY_REASON missing");
-            assert!(
-                deny_reason.contains(&format!("{action:?}")),
-                "ZEPH_DENY_REASON should contain {action:?}, got: {deny_reason}"
-            );
-        }
-    }
-
-    // --- record_shadow_event (spec 010-7 FR-001–FR-004) ---
-
-    fn make_tool_req(name: &str) -> zeph_llm::provider::ToolUseRequest {
-        zeph_llm::provider::ToolUseRequest {
-            id: format!("id_{name}"),
-            name: name.into(),
-            input: serde_json::Value::Null,
-        }
-    }
-
-    fn make_tool_call(name: &str) -> ToolCall {
-        ToolCall {
-            tool_id: zeph_common::ToolName::new(name),
-            ..Default::default()
-        }
-    }
-
-    fn make_agent_with_shadow(enabled: bool) -> Agent<crate::testing::MockChannel> {
-        use crate::testing::{MockChannel, MockToolExecutor, mock_provider};
-        use zeph_skills::registry::SkillRegistry;
-        let provider = mock_provider(vec![]);
-        let channel = MockChannel::new(vec![] as Vec<String>);
-        let registry = SkillRegistry::empty();
-        let executor = MockToolExecutor::no_tools();
-        let cfg = zeph_config::ShadowMemoryConfig {
-            enabled,
-            drift_threshold: 0.01,
-            window_size: 3,
-            max_events: 50,
-        };
-        Agent::new(provider, channel, registry, None, 5, executor).with_shadow_memory_config(&cfg)
-    }
-
-    #[test]
-    fn record_shadow_event_noop_when_disabled() {
-        let mut agent = make_agent_with_shadow(false);
-        agent.runtime.debug.iteration_counter = 1;
-        let calls = vec![make_tool_req("shell")];
-        // Must not panic; shadow_memory stays None.
-        agent.record_shadow_event(&calls, "goal".into());
-        assert!(
-            agent.services.security.shadow_memory.is_none(),
-            "shadow_memory must remain None when disabled"
-        );
-    }
-
-    #[test]
-    fn record_shadow_event_appends_event_when_enabled() {
-        let mut agent = make_agent_with_shadow(true);
-        agent.runtime.debug.iteration_counter = 1;
-        let calls = vec![make_tool_req("shell"), make_tool_req("web_scrape")];
-        agent.record_shadow_event(&calls, "test goal".into());
-        let mem = agent.services.security.shadow_memory.as_ref().unwrap();
-        assert_eq!(mem.len(), 1, "one event must be recorded after one batch");
-    }
-
-    #[test]
-    fn record_shadow_event_goal_drift_emits_security_event() {
-        use tokio::sync::watch;
-        let rt = tokio::runtime::Builder::new_current_thread()
-            .build()
-            .unwrap();
-        rt.block_on(async {
-            use crate::testing::{MockChannel, MockToolExecutor, mock_provider};
-            use zeph_skills::registry::SkillRegistry;
-            let (tx, rx) = watch::channel(crate::metrics::MetricsSnapshot::default());
-            let cfg = zeph_config::ShadowMemoryConfig {
-                enabled: true,
-                drift_threshold: 0.01,
-                window_size: 3,
-                max_events: 100,
-            };
-            let mut agent = Agent::new(
-                mock_provider(vec![]),
-                MockChannel::new(vec![] as Vec<String>),
-                SkillRegistry::empty(),
-                None,
-                5,
-                MockToolExecutor::no_tools(),
-            )
-            .with_shadow_memory_config(&cfg)
-            .with_metrics(tx);
-
-            agent.runtime.debug.iteration_counter = 1;
-
-            // Fill initial window with low-variance events.
-            let low = vec![make_tool_req("read")];
-            for _ in 0..5 {
-                agent.record_shadow_event(&low, "read files".into());
-            }
-            // Introduce high-privilege divergent batch to spike drift.
-            let high = vec![
-                make_tool_req("shell"),
-                make_tool_req("fetch"),
-                make_tool_req("write"),
-            ];
-            for _ in 0..5 {
-                agent.record_shadow_event(&high, "exfiltrate everything".into());
-            }
-
-            let snap = rx.borrow().clone();
-            // The test verifies that if GoalDrift fires, the event has the right category.
-            // (Whether it fires depends on drift score internals; we assert structural correctness.)
-            for ev in &snap.security_events {
-                if ev.category == zeph_common::SecurityEventCategory::GoalDrift {
-                    assert_eq!(ev.source, "shadow_memory");
-                    return;
-                }
-            }
-            // If no GoalDrift was emitted, at minimum confirm events were recorded.
-            let mem = agent.services.security.shadow_memory.as_ref().unwrap();
-            assert!(!mem.is_empty(), "shadow_memory must have recorded events");
-        });
-    }
-
     // Gap 3: handle_request_compaction must return the rate-limit error when
     // CompactionState is already CompactedThisTurn.
     #[test]
@@ -3703,186 +2645,6 @@ mod tests {
                 "rate-limit guard must fire: got {result:?}"
             );
         });
-    }
-
-    // Gap 5: apply_acon_compression must be a no-op when acon_config.enabled = false.
-    #[test]
-    fn apply_acon_compression_noop_when_disabled() {
-        use crate::testing::{MockChannel, MockToolExecutor, mock_provider};
-        use zeph_llm::provider::MessagePart;
-        use zeph_skills::registry::SkillRegistry;
-
-        let mut agent = Agent::new(
-            mock_provider(vec![]),
-            MockChannel::new(vec![] as Vec<String>),
-            SkillRegistry::empty(),
-            None,
-            5,
-            MockToolExecutor::no_tools(),
-        );
-        // Disable Acon.
-        agent.services.memory.subsystems.acon_config.enabled = false;
-
-        // Build a result part with text that would be truncated if Acon were active.
-        let big_content = "word ".repeat(5000);
-        let mut parts = vec![MessagePart::ToolResult {
-            tool_use_id: "id_shell".to_owned(),
-            content: big_content.clone(),
-            is_error: false,
-        }];
-        let calls = vec![make_tool_req("shell")];
-
-        agent.apply_acon_compression(&calls, &mut parts);
-
-        // Content must be unchanged.
-        if let MessagePart::ToolResult { content, .. } = &parts[0] {
-            assert_eq!(
-                content.len(),
-                big_content.len(),
-                "content must not be modified when acon is disabled"
-            );
-        } else {
-            panic!("expected ToolResult part");
-        }
-    }
-
-    // spec-072 C5/AC-15 (T-213): pre-assembly pass safety with an interleaved Image sibling.
-    // `run_causal_ipi_post_probe` and `record_shadow_event` take `result_parts`/`tool_calls`
-    // by shared reference and never touch `MessagePart::Image` at all, so they cannot
-    // mutate/drop it by construction. `apply_acon_compression` is the only pass that mutates
-    // `result_parts` in place — this test proves its `tool_use_id`-based `ToolResult`
-    // targeting is unaffected by the presence/position of a non-`ToolResult` sibling, and
-    // that the `Image` part itself survives all three passes byte-for-byte.
-    #[test]
-    #[allow(clippy::too_many_lines)] // control + interleaved runs, both passes asserted
-    fn pre_assembly_passes_preserve_image_sibling() {
-        use crate::testing::{MockChannel, MockToolExecutor, mock_provider};
-        use zeph_llm::provider::{ImageData, ToolUseRequest};
-        use zeph_skills::registry::SkillRegistry;
-
-        fn make_agent() -> Agent<MockChannel> {
-            let mut agent = Agent::new(
-                mock_provider(vec![]),
-                MockChannel::new(vec![] as Vec<String>),
-                SkillRegistry::empty(),
-                None,
-                5,
-                MockToolExecutor::no_tools(),
-            );
-            // Default passthrough_threshold (2000 tokens) is well below the ~9000-token
-            // bodies below, so compression actually runs (not a PassThrough no-op).
-            agent.services.memory.subsystems.acon_config.enabled = true;
-            agent
-        }
-
-        fn tool_result(id: &str, content: String) -> MessagePart {
-            MessagePart::ToolResult {
-                tool_use_id: id.to_owned(),
-                content,
-                is_error: false,
-            }
-        }
-
-        let big_a = "alpha ".repeat(3000);
-        let big_b = "bravo ".repeat(3000);
-        let calls = vec![
-            ToolUseRequest {
-                id: "id_a".to_owned(),
-                name: "read".into(),
-                input: serde_json::Value::Null,
-            },
-            ToolUseRequest {
-                id: "id_b".to_owned(),
-                name: "read".into(),
-                input: serde_json::Value::Null,
-            },
-        ];
-        let image_bytes = vec![1u8, 2, 3, 4, 5];
-        let image_mime = "image/png".to_owned();
-        let image = MessagePart::Image(Box::new(ImageData {
-            data: image_bytes.clone(),
-            mime_type: image_mime.clone(),
-        }));
-
-        // Control run: no Image sibling at all.
-        let big_a_original_len = big_a.len();
-        let mut control_parts = vec![
-            tool_result("id_a", big_a.clone()),
-            tool_result("id_b", big_b.clone()),
-        ];
-        let mut control_agent = make_agent();
-        control_agent.apply_acon_compression(&calls, &mut control_parts);
-
-        // Interleaved run: Image positioned between the two ToolResult parts.
-        let mut interleaved_parts = vec![
-            tool_result("id_a", big_a),
-            image,
-            tool_result("id_b", big_b),
-        ];
-        let mut agent = make_agent();
-        let rt = tokio::runtime::Builder::new_current_thread()
-            .build()
-            .unwrap();
-        rt.block_on(async {
-            agent
-                .run_causal_ipi_post_probe(None, &interleaved_parts)
-                .await;
-        });
-        agent.record_shadow_event(&calls, "goal summary".into());
-        agent.apply_acon_compression(&calls, &mut interleaved_parts);
-
-        // (a) Compression output for both ToolResult parts is unaffected by the interleaved
-        // Image sibling: identical to the control run without it.
-        let MessagePart::ToolResult {
-            content: control_a, ..
-        } = &control_parts[0]
-        else {
-            panic!("expected ToolResult part in control run");
-        };
-        let MessagePart::ToolResult {
-            content: control_b, ..
-        } = &control_parts[1]
-        else {
-            panic!("expected ToolResult part in control run");
-        };
-        let MessagePart::ToolResult {
-            content: interleaved_a,
-            ..
-        } = &interleaved_parts[0]
-        else {
-            panic!("expected ToolResult part at index 0");
-        };
-        let MessagePart::ToolResult {
-            content: interleaved_b,
-            ..
-        } = &interleaved_parts[2]
-        else {
-            panic!("expected ToolResult part at index 2");
-        };
-        assert!(
-            control_a.len() < big_a_original_len,
-            "sanity: compression must actually run (control_a shorter than original)"
-        );
-        assert_eq!(
-            control_a, interleaved_a,
-            "id_a compression must be identical with/without the interleaved Image sibling"
-        );
-        assert_eq!(
-            control_b, interleaved_b,
-            "id_b compression must be identical with/without the interleaved Image sibling"
-        );
-
-        // (b) The Image part itself survives all three passes byte-for-byte.
-        match &interleaved_parts[1] {
-            MessagePart::Image(img) => {
-                assert_eq!(img.data, image_bytes, "Image bytes must be unchanged");
-                assert_eq!(
-                    img.mime_type, image_mime,
-                    "Image mime_type must be unchanged"
-                );
-            }
-            other => panic!("expected Image part at index 1, got {other:?}"),
-        }
     }
 
     // Regression guard for #5584: when a retryable failure (e.g. Qdrant unreachable) was
@@ -4100,440 +2862,6 @@ mod tests {
             "must not mandate yet another detour, got: {}",
             hints[0]
         );
-    }
-
-    mod reformat_phase_tests {
-        use zeph_tools::registry::{InvocationHint, ToolDef};
-
-        use super::*;
-        use crate::agent::agent_tests::*;
-
-        fn test_tool_def() -> ToolDef {
-            ToolDef {
-                id: "test_tool".into(),
-                description: "a test tool".into(),
-                schema: schemars::Schema::default(),
-                invocation: InvocationHint::ToolCall,
-                output_schema: None,
-                server_id: None,
-            }
-        }
-
-        fn invalid_params_error() -> zeph_tools::ToolError {
-            zeph_tools::ToolError::InvalidParams {
-                message: "path must be a string".to_owned(),
-            }
-        }
-
-        fn bad_tool_call() -> ToolCall {
-            let mut params = serde_json::Map::new();
-            params.insert("path".to_owned(), serde_json::json!(123));
-            ToolCall {
-                tool_id: zeph_common::ToolName::new("test_tool"),
-                params,
-                ..Default::default()
-            }
-        }
-
-        fn tool_use_request() -> zeph_llm::provider::ToolUseRequest {
-            zeph_llm::provider::ToolUseRequest {
-                id: "call-1".to_owned(),
-                name: "test_tool".to_owned().into(),
-                input: serde_json::json!({"path": 123}),
-            }
-        }
-
-        #[tokio::test]
-        async fn retries_with_corrected_arguments_on_success() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor = MockToolExecutor::new(vec![Ok(Some(zeph_tools::ToolOutput {
-                tool_name: "test_tool".to_owned().into(),
-                summary: "done".to_owned(),
-                blocks_executed: 1,
-                filter_stats: None,
-                diff: None,
-                streamed: false,
-                terminal_id: None,
-                locations: None,
-                raw_response: None,
-                claim_source: None,
-                ..Default::default()
-            }))])
-            .with_definitions(vec![test_tool_def()]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            agent.tool_orchestrator.parameter_reformat_provider = "fast".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            let output = tool_results
-                .remove(0)
-                .expect("reformat retry should succeed")
-                .expect("tool output should be present");
-            assert_eq!(output.summary, "done");
-        }
-
-        #[tokio::test]
-        async fn keeps_original_error_when_provider_returns_malformed_json() {
-            let provider = mock_provider(vec!["not json at all".into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor =
-                MockToolExecutor::new(vec![Ok(None)]).with_definitions(vec![test_tool_def()]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            agent.tool_orchestrator.parameter_reformat_provider = "fast".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                matches!(
-                    tool_results[0],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "original error must be preserved on parse failure"
-            );
-        }
-
-        #[tokio::test]
-        async fn keeps_original_error_when_tool_schema_unknown() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            // No `with_definitions` — schema lookup for "test_tool" fails.
-            let executor = MockToolExecutor::new(vec![Ok(None)]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            agent.tool_orchestrator.parameter_reformat_provider = "fast".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                matches!(
-                    tool_results[0],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "original error must be preserved when schema is unknown"
-            );
-        }
-
-        #[tokio::test]
-        async fn is_a_noop_when_provider_not_configured() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor =
-                MockToolExecutor::new(vec![Ok(None)]).with_definitions(vec![test_tool_def()]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            // parameter_reformat_provider left empty (default) — FR: disabled means no LLM call.
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                matches!(
-                    tool_results[0],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "reformat must not run when parameter_reformat_provider is empty"
-            );
-        }
-
-        // Regression test: when the provider-pool registry was never wired for this `Agent`
-        // (empty `provider_pool`, no `provider_config_snapshot` — the state of a lightweight
-        // test/bootstrap agent that never called `with_provider_pool`, never a real production
-        // agent since `validate_pool` rejects an empty `[[llm.providers]]` list at config-load
-        // time), `reformat_tool_call` still falls back to the primary provider, matching every
-        // other `resolve_background_provider` call site (e.g. `compress_provider`). Only once
-        // the registry IS wired does an unresolvable name become a real misconfiguration (see
-        // `is_a_noop_when_registry_wired_but_name_absent_from_pool` below).
-        #[tokio::test]
-        async fn falls_back_to_primary_when_registry_never_wired() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor = MockToolExecutor::new(vec![Ok(Some(zeph_tools::ToolOutput {
-                tool_name: "test_tool".to_owned().into(),
-                summary: "done".to_owned(),
-                blocks_executed: 1,
-                filter_stats: None,
-                diff: None,
-                streamed: false,
-                terminal_id: None,
-                locations: None,
-                raw_response: None,
-                claim_source: None,
-                ..Default::default()
-            }))])
-            .with_definitions(vec![test_tool_def()]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            // Non-empty name, but provider_pool is empty — name cannot resolve.
-            agent.tool_orchestrator.parameter_reformat_provider = "unregistered".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            let output = tool_results
-                .remove(0)
-                .expect("reformat should run using the primary provider fallback")
-                .expect("tool output should be present");
-            assert_eq!(output.summary, "done");
-        }
-
-        /// Minimal `ProviderConfigSnapshot` fixture shared by the registry-wired regression
-        /// tests below — all `None`/empty since the fields under test don't need real secrets.
-        fn empty_snapshot() -> crate::agent::state::ProviderConfigSnapshot {
-            crate::agent::state::ProviderConfigSnapshot {
-                claude_api_key: None,
-                openai_api_key: None,
-                gemini_api_key: None,
-                compatible_api_keys: std::collections::HashMap::new(),
-                llm_request_timeout_secs: 30,
-                embedding_model: String::new(),
-                gonka_private_key: None,
-                gonka_address: None,
-                cocoon_access_hash: None,
-            }
-        }
-
-        // Regression test for #5478: once the provider-pool registry IS wired (non-empty
-        // `provider_pool`, the state every real production `Agent` is in per #5450), a
-        // configured `parameter_reformat_provider` name that does not match any entry is a real
-        // misconfiguration and must no-op (keep the original tool error) rather than silently
-        // reformatting with the primary provider.
-        #[tokio::test]
-        async fn is_a_noop_when_registry_wired_but_name_absent_from_pool() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor =
-                MockToolExecutor::new(vec![Ok(None)]).with_definitions(vec![test_tool_def()]);
-
-            // The pool is wired (non-empty), but registers a different name than the one
-            // configured below, so "unregistered" still cannot resolve.
-            let other_entry = crate::config::ProviderEntry {
-                provider_type: crate::config::ProviderKind::Ollama,
-                name: Some("other".into()),
-                ..Default::default()
-            };
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor)
-                .with_provider_pool(vec![other_entry], empty_snapshot());
-            agent.tool_orchestrator.parameter_reformat_provider = "unregistered".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                matches!(
-                    tool_results[0],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "reformat must no-op when the registry is wired but the configured name is \
-                 absent from the pool, not silently fall back to the primary provider"
-            );
-        }
-
-        // Regression test for #5600: a configured `parameter_reformat_provider` name that IS
-        // present in `provider_pool` but whose provider construction fails (e.g. a required
-        // secret is missing from the config snapshot) must no-op — unlike the registry-not-wired
-        // case above, this must not silently fall back to the primary provider.
-        #[tokio::test]
-        async fn is_a_noop_when_registry_wired_but_provider_build_fails() {
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/corrected"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor =
-                MockToolExecutor::new(vec![Ok(None)]).with_definitions(vec![test_tool_def()]);
-
-            // "broken" is present in provider_pool, but is a Claude entry with no API key in
-            // the snapshot, so `build_provider_for_switch` fails at resolve time.
-            let broken_entry = crate::config::ProviderEntry {
-                provider_type: crate::config::ProviderKind::Claude,
-                name: Some("broken".into()),
-                ..Default::default()
-            };
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor)
-                .with_provider_pool(vec![broken_entry], empty_snapshot());
-            agent.tool_orchestrator.parameter_reformat_provider = "broken".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                matches!(
-                    tool_results[0],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "reformat must no-op when the in-pool provider fails to build, not silently \
-                 fall back to the primary provider"
-            );
-        }
-
-        #[tokio::test]
-        async fn replaces_original_error_when_retry_still_fails() {
-            // FR-004: the retried outcome MUST replace tool_results[idx] even when the retry
-            // itself fails — the reformat phase gives up cleanly after a single attempt rather
-            // than looping, and never leaves the pre-reformat error in place.
-            let provider = mock_provider(vec![r#"{"arguments":{"path":"/still-bad"}}"#.into()]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor = MockToolExecutor::new(vec![Err(zeph_tools::ToolError::InvalidParams {
-                message: "still not a valid path".to_owned(),
-            })])
-            .with_definitions(vec![test_tool_def()]);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            agent.tool_orchestrator.parameter_reformat_provider = "fast".to_owned();
-
-            let tool_calls = vec![tool_use_request()];
-            let calls = vec![bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            match &tool_results[0] {
-                Err(zeph_tools::ToolError::InvalidParams { message }) => {
-                    assert_eq!(
-                        message, "still not a valid path",
-                        "the retried failure must replace the original error, not leave the \
-                         pre-reformat error in place"
-                    );
-                }
-                other => {
-                    panic!("expected the retried failure to replace the original, got {other:?}")
-                }
-            }
-        }
-
-        #[tokio::test]
-        async fn budget_exhausted_skips_remaining_calls_in_same_phase() {
-            // Regression test for the timing bug: `reformat_start` used to be recreated
-            // immediately before each per-call elapsed check, making the budget guard an
-            // effective no-op. It is now created once before the loop, so real time consumed
-            // by an earlier reformat call in the same phase counts against later calls' budget.
-            let provider = mock_provider(vec![
-                r#"{"arguments":{"path":"/corrected"}}"#.into(),
-                r#"{"arguments":{"path":"/corrected"}}"#.into(),
-            ]);
-            let channel = MockChannel::new(vec![]);
-            let registry = create_test_registry();
-            let executor = MockToolExecutor::new(vec![Ok(Some(zeph_tools::ToolOutput {
-                tool_name: "test_tool".to_owned().into(),
-                summary: "done".to_owned(),
-                blocks_executed: 1,
-                filter_stats: None,
-                diff: None,
-                streamed: false,
-                terminal_id: None,
-                locations: None,
-                raw_response: None,
-                claim_source: None,
-                ..Default::default()
-            }))])
-            .with_definitions(vec![test_tool_def()])
-            // Consumes >1s of wall time on the first (only) dispatched retry, so the second
-            // tool call's budget check — using the same `reformat_start` — sees the whole-phase
-            // budget already exhausted.
-            .with_delay(1_100);
-            let mut agent = Agent::new(provider, channel, registry, None, 5, executor);
-            agent.tool_orchestrator.parameter_reformat_provider = "fast".to_owned();
-            agent.tool_orchestrator.max_retry_duration_secs = 1;
-
-            let tool_calls = vec![tool_use_request(), tool_use_request()];
-            let calls = vec![bad_tool_call(), bad_tool_call()];
-            let mut tool_results: Vec<
-                Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-            > = vec![Err(invalid_params_error()), Err(invalid_params_error())];
-            let cancel = tokio_util::sync::CancellationToken::new();
-
-            agent
-                .handle_reformat_phase(&tool_calls, &calls, &mut tool_results, &cancel)
-                .await
-                .unwrap();
-
-            assert!(
-                tool_results[0].is_ok(),
-                "first call is within budget and should be reformatted successfully"
-            );
-            assert!(
-                matches!(
-                    tool_results[1],
-                    Err(zeph_tools::ToolError::InvalidParams { .. })
-                ),
-                "second call must be skipped once the whole-phase budget is exhausted by the \
-                 first call's real elapsed time"
-            );
-        }
     }
 
     // --- PAAC tool-dispatch unmasking: unmask_json_value (#5437) ---
